requirements.txt
Django~=5.1.4
django-tinymce~=4.1.0